refactor:restructure folder hierarchy
