+++ src/main/java/com/tenbit/beep/attendance/controller/AttendanceController.java
@@ -1,6 +1,5 @@
 package com.tenbit.beep.attendance.controller;
 
-import com.tenbit.beep.attendance.service.AttendanceServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AttendanceController {
 
-    private final AttendanceServiceImpl attendanceService;
+    private final AttendanceService attendanceService;
 
     @PostMapping("/check")
     public ResponseEntity<String> checkAttendance(@RequestParam String publicId) {
